Rename components for accuracy and move data-consuming components to containers
