+++ src/apiDefs/verification.ts
@@ -1,7 +1,7 @@
-import { IApiDescription } from './schema';
+import { IApiDefinition } from './schema';
 
 const swaggerHost : string = process.env.REACT_APP_VETSGOV_SWAGGER_API!;
-const verificationApis : IApiDescription[] = [
+const verificationApis : IApiDefinition[] = [
   {
     description: "Get a Veteran's disability rating",
     docSources: [
