+++ src/components/index.ts
@@ -4,10 +4,5 @@ export * from './NavBar';
 export * from './Footer';
 export * from './PageContent';
 export * from './PageHero';
-export * from './ApiPage';
-export * from './ApiPageReleaseNotes';
 export * from './PageHeader';
 export * from './QuickstartWrapper';
-import SwaggerDocs from './SwaggerDocs';
-
-export { SwaggerDocs };
