+++ src/containers/CategoryReleaseNotesPage.tsx
@@ -4,12 +4,12 @@ import { Flag } from 'flag';
 import { RouteComponentProps } from 'react-router';
 
 import { apiDefs, IApiDescription } from '../apiDefs';
+import CardLink from '../components/CardLink';
 import PageHeader from '../components/PageHeader';
+import { VAInternalOnlyTag } from '../components/VAInternalOnlyTag';
 import { IApiNameParam } from '../types';
-import CardLink from './CardLink';
-import { VAInternalOnlyTag } from './VAInternalOnlyTag';
 
-export class ApiPageReleaseNotes extends React.Component<RouteComponentProps<IApiNameParam>, {}> {
+export default class CategoryReleaseNotesPage extends React.Component<RouteComponentProps<IApiNameParam>, {}> {
   public render() {
     const { apiCategoryKey } = this.props.match.params;
     const { apis, releaseNotes } = apiDefs[apiCategoryKey];
