+++ src/containers/swaggerPlugins/VersionSelect.scss
